♻️ refactor: Using 2D Array now
Changed to Using 2D arrays for better play with pytorch
‼️ Warning: This change is detrimental to CPU performance. Runs
significantly slower

🐛 fix: Second initialization loop should of had been range (m+1)(n+2).
Instead, the first loop had this range... which was also removed because
pytorch is already initialization all these to zeroes. smh

Also fills cells next to the border before ODE and PDE loop, meaning
that we do not have to check these after the main loops. This removes
the edge cases.